requirements.txt
python-telegram-bot==21.5
python-dotenv==1.0.1
aiosqlite==0.20.0